services/summary_of_content_changes/requirements.txt
ollama
pika
python-dotenv